requirements.txt
requests>=2.28
dnspython>=2.3.0
ratelimit>=2.2.0
